Remove pre-Go1.9 fallback code
Go 1.12beta1 is out and even Appengine's oldest runtime is Go 1.9,
so get rid of pre-1.9 fallback code and always use math/bits.
