+++ xxhash.go
@@ -4,6 +4,7 @@ package xxhash
 
 import (
 	"encoding/binary"
+	"math/bits"
 )
 
 const (
@@ -173,3 +174,12 @@ func mergeRound(acc, val uint64) uint64 {
 	acc = acc*prime1 + prime4
 	return acc
 }
+
+func rol1(x uint64) uint64  { return bits.RotateLeft64(x, 1) }
+func rol7(x uint64) uint64  { return bits.RotateLeft64(x, 7) }
+func rol11(x uint64) uint64 { return bits.RotateLeft64(x, 11) }
+func rol12(x uint64) uint64 { return bits.RotateLeft64(x, 12) }
+func rol18(x uint64) uint64 { return bits.RotateLeft64(x, 18) }
+func rol23(x uint64) uint64 { return bits.RotateLeft64(x, 23) }
+func rol27(x uint64) uint64 { return bits.RotateLeft64(x, 27) }
+func rol31(x uint64) uint64 { return bits.RotateLeft64(x, 31) }
